15 x 15 game functional with starting and game over screen
